MemPool: Remove duplicated code
The only reason for this commitN(), preallocN() and allocN having
their own code was that they used the unsigned long int, in order
to permit huge allocations.

However, nowhere in the software do we need such huge allocations,
so let's remove the duplicated code and just call the coressponding
*Bytes() functions.

+++ mm/mempool.cpp
@@ -26,7 +26,7 @@ extern "C" {
 #include <unistd.h>
 }
 
-MemPool::MemPool(unsigned int nr_pages, unsigned long objsize)
+MemPool::MemPool(unsigned int nr_pages, unsigned int objsize)
 {
 	poolSize = nr_pages * sysconf(_SC_PAGESIZE);
 	objSize = objsize;
